requirements.txt
growattServer>=2.0.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.0
pyyaml>=6.0
python-dateutil>=2.9.0
httpx>=0.27.0
